[FIX] l10n_be_reports: 281.50 pdf add national number if available
The aim of this commit is to add the national number on the 281.50 pdf
if it is available.

Context:
The customer expects it to be filled if they filled it and it helps the
authorities to identify the receiver.

Before this commit:
The national number wasn't added even if we had it.

After this commit:
The national number is added to the pdf if it is available for that
partner.

opw-4852895

closes odoo/enterprise#87255

X-original-commit: 13321c19a190f8ff8c3e12886a5106cdce78917d
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>