[IMP] esg: add an end date to other emissions
Some emissions can represent activities that occur over a period of time.
One of such emissions is the employee commuting, which needs to be added in a later commit.
This commit therefore adds an optional end date to emissions (other emissions only, not accounting emissions of course).

Task-4596253

Part-of: odoo/enterprise#87519
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>