[FIX] account_accountant: _retrieve_partner domain fix
There was a corner case where a single partner should match a statement line but would not.
The case was when the statement line account number matches multiple accounts
that are linked to the same active partner.
This commit fixes this issue by relying on the number of partners instead of the number of banks
while matching.

task-4826477

closes odoo/enterprise#87246

X-original-commit: 0d4625367921cdd7752734198d20e9f3678d06ba
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>