[IMP] ai: implementation of an audio transcription component
This commit introduces a new transcription component that can be
inserted in the html editor. This component allows the user to start a
recording of his microphone. The transcription is made in real time in
real time. When the recording is stopped, a summary of the transcription
is added. The component also allows the user to input notes manually.

task-4791084

closes odoo/enterprise#87615

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>