[FIX] account_reports: fix traceback when user add expression label in column
Currently a raceback occurs when the user changes/adds the expression label
of the column in the "Generic Tax report".

To reproduce this issue:
1) Install "account_reports"
2) Open "Generic Tax report" from "Accounting Reports"
3) In columns change the "Expression Label" and save the record
4) Open the "Tax Report" in "Reporting"
5) A traceback occurs

Error:-
"UnboundLocalError: local variable 'col_value' referenced before assignment"

Cause:
The "col_value" is assigned based on the "expr_label" if it doesn't
match the if conditions it leads to a traceback.
https://github.com/odoo/enterprise/blob/d154cbf1bd5b4cc104ff0e2443047aff0c05330f/account_reports/models/account_generic_tax_report.py#L920-L932

Solution:
This commit will resolve this issue by assigning a fallback value
of an empty string to col_value.

opw-4841789

closes odoo/enterprise#87581

X-original-commit: 1203a8b31d2d5376ebf46ced83eb12128b03a502
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>