[FIX] sale_purchase_stock_inter_company_rules : Set Correct Warehouse
Issue:
--------
While having multiple companies, multiple warehouses and when different
warehouses are set for different companies other than the first one(id=1)
under Inter-Company Transactions "Use Warehouse".
Now, when we check the 'Generate Purchase Orders' then the 'Use Warehouse'
value which was set before  is getting overridden to the first Warehouse
which has minimal 'id'.

Cause:
------
[Here](https://github.com/odoo/enterprise/commit/6324a6bab04fa0f3aa6d87deabc44c7c2eafc295#diff-51563a07f4b65f4ffe54bfb161ac1e9e58ff0f0bcf2522c9f6aa3746b60a68adR28-R38)
Since, whenever there's a change within any of those check-boxes the compute
will be triggered and the values will get modified. During this trigger the
value set in the 'Use Warehouse' is getting overridden to the warehouse
with minimal 'id'(For ex: id-1).

Solution:
-----------
To fix this, we'll just check if there's no value set before. If yes, then we'll
 set the first warehouse which has minimal 'id'. Otherwise, set the one
which is selected.

Steps to reproduce:
-------------------------
1. Create a fresh db with multiple warehouses and
'sale_purchase_stock_inter_company_rules' module installed.
2. Go to General Settings> Inter-Company Transactions and check/select
'Generate Sale Orders' and set a different warehouse and save.
3. Now check/select the 'Generate Purchase Orders' and save.
4. Look at the Use Warehouse value in the Settings. It will be the first one
with minimal 'id'.

Ref PRs:
1. https://github.com/odoo/enterprise/pull/55350

closes odoo/enterprise#99878

X-original-commit: 12cfc4dd1335332c9ec7394d4f91d24bffd3645b
Signed-off-by: Steve Van Essche <svs@odoo.com>