[FIX] l10n_pe_edi_pos: add method to check if order is invoiced
Steps to reproduce
- install `l10n_pe_edi_pos` module
- open a pos session
- create a order, when checking out, tick `invoice` checkbox
- pay and complete the order
- go to order menu, (present in navbar)
- select the previously completed order, apply for refund
Observation: you are navigated to Order screen
Expectation: A dialog for reason should be open first

Issue
- after this commit odoo/odoo@3a09ea7 and odoo/enterprise@e3e84bb account_move
field is used instead of `invoiced` state of order. But the `account_move` field
is always undefined.
This is due to fact that when pos_order is fulfilled from `syncAllOrders`
function, a call to `missingRecursive` is made, as `account.move` model is never
 loaded, it is unable to link it to appropriate model and results to undefined

Fix
- use order.raw.account_move instead of order.account_move

opw-4744785

closes odoo/enterprise#90274

X-original-commit: a85228539db6e4b79ba7ee304d179ccb1020a1db
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Ajit Kumar  Krishnanand Singh (aksi) <aksi@odoo.com>