[FIX] esg: prevent non-expense accounts creation and clean-up lines
Before this commit, it was possible to create accounts from the assignation lines list view of another type than "Expense" or "Fixed Assets".
This is not desired, as this type of account is not part of the domain of the 'account_id' field of the assignation line model.
We prevent creation of such accounts from ESG in general, to avoid confusion and ensure data integrity.

Moreover, when changing the account type of an existing account, if this one is changed to a type other than "Expense" or "Fixed Assets",
we verify if there are some assignation lines linked to this account and/or journal items linked to that account.
If so, we raise a warning to the user and if he wants to proceed, we remove the assignation lines linked to this account and remove the emission factors of journal entries linked to that account.

task-4859806

closes odoo/enterprise#95833

X-original-commit: 92793fb57722e65aa10feff6e67adaecb4ed1ce5
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Maxime de Neuville (mane) <mane@odoo.com>