[FIX] l10n_ae_corporate_tax_report: search syntax
When generating the accounting entry from the Corporate Tax Report, the
system first tries to find a miscellaneous journal with the short code
`MISC`.

If this specific journal does not exist, the logic attempts to fall back
and find any other journal of type 'Miscellaneous' (`type = 'general'`).
However, this fallback search fails with an error due to an incorrect
search domain syntax.

Steps to reproduce:
1. Go to Accounting > Configuration > Journals.
2. Ensure no journal exists with the short code MISC.
3. Ensure at least one other journal of type "Miscellaneous" exists
    (for example, with short code GEN or MSC).
4. Go to Accounting > Reporting > Corporate Tax Report.
5. Try to generate the accounting entry.

opw-4836686

closes odoo/enterprise#87794

X-original-commit: 5d10fb2c734d1379c1415f683f3c5ab0620d425b
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Corentin Thaon (thco) <thco@odoo.com>