[FIX] web_studio: fix undeterminism
Before this commit there was an undterminism because we asserted imperatively some inputs had some values.

With the advent of Hoot and its pseudo-selectors in tours it is now possible to have tour steps that have a trigger
that in fact assert some stuff declaratively.

After this commit, at least this source of undeterminisms in that tour is lifted.

build_error-163405

closes odoo/enterprise#88584

X-original-commit: 0081b4740967db25ede908c5ef428335ee9f1dc7
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Omar Sherif Ali Hassan (osah) <osah@odoo.com>