[IMP] helpdesk: allow editing stage in list view based on team consistency
Before this commit:
The stage field was always readonly in the list view, preventing users from
changing the stage of a ticket, whether selecting a single record or multiple
records.

After this commit:
The stage field becomes editable in the list view if all selected tickets belong
to the same team. If the selection includes tickets from different teams, the
field will be readonly to prevent inconsistent stage updates.

task-4775290

closes odoo/enterprise#87587

Related: odoo/odoo#215419
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>