[FIX] voip: fix falsy `t9_name` when the partner name is `False`.
This commit fixes a bug that happens when the partner name is a falsy
value like a `" "`. This made the `t9_name` to be `false` and then to
give an error when trying to access the `trim()` function.

A solution for this was to guard it with a condition to check if the
t9_name is falsy or not.

Reproduction steps of the bug:
1. Create a contact where their name is a space (`" "`).
2. Include a phone number for this contact.
3. Open the VoIP wizard then the keypad tab.
4. try to search for that contact using the keypad numbers.
5. you will get the error saying that `trim()` is not a function.

closes odoo/enterprise#87560

X-original-commit: 614b509b79ab36cdc4c4aa6bf2404fc4ed91e952
Signed-off-by: Yuchen Huang (yhu) <yhu@odoo.com>
Signed-off-by: Mohammed Basioni (basm) <basm@odoo.com>