[IMP] account_reports,l10n_*_reports: Simplify return instantiation
Previously, we had to define the return type in data and also
instanciate it in Python.

Now, we can specify the country_code when defining the return
type which will allow it to be created in the base function
generate_all_returns, if it's not specified it can still be
created by hand with a custom override of the function.

task-4841751

closes odoo/enterprise#87329

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>