[FIX] l10n_uk_reports: remove garbage value from localStorage
There are still Odoo requests that are sent to hmrc with invalid
'Gov-Client-Device-ID' header.  They are showing this error:
"Submit a UUID which is 128 bits or 32 hex characters long".

A possible explanation, is that some users have some garbage value
in the localStorage for 'hmrc_gov_client_device_id', that does not
correspond to a uuid.  This value would then be sent each time in the
headers, and get rejected.

The fix here is to clear the localStorage value if it is not a uuid.

task-4627086

closes odoo/enterprise#96535

X-original-commit: 6c8aeef1fed0ec7c352a8d7874d1396c565eef00
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Arnaud Sibille (arsi) <arsi@odoo.com>