[FIX] knowledge: show access errror for embedded view
When sharing a knowledge article with an embedded view
to an external user. The view was loading indefinitely
without saying that the user doesn't have access to that view.

Steps to reproduce:
-------------------
* Go to the Planning app
* Click the 'gear' icon
* Click 'Knowledge'
* Click 'Inset view in article'
* Select an article to add it to
* Share the article with portal user

> Observation:
Embedded planning keeps loading indefinitely.
Works with other views that require access rights.

Why the fix:
------------
Embedded view with loading regardless of the error state.

opw-4664036

closes odoo/enterprise#88847

X-original-commit: 7e512c45fa2967f40f65f0de1d49c9ab32a9853c
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Guillaume Vardakis (gvar) <gvar@odoo.com>