[IMP] appointment: add active field on appointment.question
In order to give more flexibility to the appointment admin,
allow to archive questions. This will hide them in the modal
of question selection when adding one on an Event.

Also, this allows not removing all existing linked answers /
answer inputs as it is the case at the moment by archiving
the record instead of deleting it, by switching to a 'restrict'
ondelete on the appointment.answer.input model.

Add a ribbon on the form to emphasize the archived status.

Task-4655040

Part-of: odoo/enterprise#82973
Related: odoo/upgrade#7691
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>