[FIX] account_accountant: limit search view dialog
Before this commit, when entering the search view dialog a searchRead was done
to create a dict of displayed move line per id. This dict is useful to compute
the residual amount when selecting a line. The problem is that we don't have any
limit on the search which is bad perf wise.

Now we modified the selection changed to give us the selected lines with the
data needed to compute the remaining amount.

Special case when we do a select domain, since the model.root.record will not
have the records. So we need to do the webRead on that.

no task id

closes odoo/enterprise#87964

X-original-commit: 74306d39479f9cfb24bb749a520c533df3f7e404
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>