[FIX] account_accountant: Cancel entries bank rec widget
During this commit: https://github.com/odoo/enterprise/commit/0931f83a37a66fbc523fa40cc97f1cc52443bd2f
we removed the domain on cancel entries to add tags on the statement itself and
a filter. But when doing the refactor of the bank rec widget, we decided to
remove the tags for now but didn't put back the domain.
This commit will add it back

no task id

closes odoo/enterprise#87454

X-original-commit: 004b61c268519131662cf625e2bd55e42f544313
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>