[FIX] knowledge: prevent discarding changes when moving an article
Currently, when a user moves another article beneath the one they are
editing, the record is reloaded, causing any unsaved changes to be lost.
This can be especially frustrating if the user has made significant edits.

To address this issue, we will ensure that any changes made to the
article are saved before the record is reloaded. This will prevent
accidental loss of work and improve the editing experience.

Steps to reproduce the issue:
1. Open an article
2. Make changes to the article (e.g, type or edit text)
3. Drag and drop another article beneath the current one.
=> The article reloads, and your unsaved changes are lost.

In the `moveArticle` function, we will check if the record has unsaved
changes (i.e., if it's "dirty"). If so, we will save the current content
before triggering a reload via `this.props.record.model.load()`.

Task-4636494

closes odoo/enterprise#93103

X-original-commit: 9a052a0972aca09fee92147d6d5f15308c799d98
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>