[FIX] web_map: compute google map urls using full addresses
**Steps to reproduce:**
- Install Contact app
- Create a contact with a specific address
- Go to the Map View of the Contact app
- Filter to view the new contact
- Position in the map might be slightly different from given one (when using OpenStreeMap)
- Click on the position marker > `Navigate To` the address is recomputed correctly
- Click on `View in Google Maps` the address given is often wrong

**Issue:**
Previous solution was trying to build the url used by the `View in Google Maps` button by using
`partner_latitude` and `partner_longitude`. These were previously computed using the default
geolocalization method. If it was set on OpenStreetMap, the coordinates were not precise
enough and impacted the Google Maps results.

As described in the documentation : `OpenStreetMap might not always be accurate.`
But this shouldn't impact Google Place API results.

**Fix:**
Adapted the computation of `googleMapUrl()` to use `contact_address_complete` to ensure the
addresses are recomputed properly when sent to Google Maps.

opw-4649910

closes odoo/enterprise#89730

X-original-commit: 20597451710d1affc24e5b16ebfa4d1432beb70a
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>
Signed-off-by: Florentin Delcourt (defl) <defl@odoo.com>