[IMP] spreadsheet_edition: unify date global filters
This commit adapts the enterprise codebase to the changes made in the
community regarding the date global filters.

This includes the removal of the `select` component, which was used
to select the date type in the date filter editor. As it's no longer
needed, it has been removed.

closes odoo/enterprise#87179

Task: 4822857
Related: odoo/odoo#213323
Related: odoo/upgrade#7822
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>