[IMP] *hr_payroll*: Refactor payslips flows with issues
===

Refactor payrun flows / possible states:
* we no longer distinguish between 'draft' and 'verify' payslips
* possible states for payslip have been changed as follows:
    * draft; verify -> draft (1)
    * done -> validated
    * paid (unchanged)
    * cancel (unchanged)
* possible states for payruns have been changed as follows:
    * draft; verify -> ready
    * close ("Done") (unchanged)
    * cancel (unchanged)

(1)
A draft payslip use to be a payslip before the sheet computation
Since payslips are supposed to be generated mostly as parts of payruns,
which run a computation at creation, this state has limited use.
When we really need to distinguish the 2 cases we now check whether the
payslip has any lines

===

Transform popup erros into 'issue's where it makes sense.
An 'issue' is a warning that appears on a payslip and that can either be
blocking (in which case the payslip state cannot go forward, and it
will appear red)
or not (it will appear yellow/orange)

task-4720075

Part-of: odoo/enterprise#87977
Related: odoo/upgrade#7958
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>