[IMP] esg_hr_fleet: create emissions from employee commuting
We currently have reports to display the estimated emissions from the employee commuting using the fleet app.
However, there is no easy way to create emitted emissions from that data.
With this commit, we ad a new button in the pivot report of the employee commuting, to allow users to create
new emissions between chosen dates.
We also display a warning in case part of the selected date range is already covered by generated emissions.

Task-4596253

closes odoo/enterprise#87519

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>