[IMP] stock_barcode: Filter operations on packaging barcode
Adds the ability to filter transfers with a packaging barcode.

BEFORE: Can only filter operations in the barcode app based
on product, packages or lot_id

NOW: Users can now filter moves with packaging barcode,
returning moves that contain products using that packaging.

WHY: Because sometimes only the packaging barcode is
accessible and not the product's

*We do not want to filter moves with the packaging matching
the packaging barcode, as sometimes packaging on the move
is not the same as source (eg. confirming PO for one pack of 6
-> move packaging will be units -> still want to find it)

task 4729518

Part-of: odoo/enterprise#87450
Related: odoo/odoo#213852
Signed-off-by: Steve Van Essche <svs@odoo.com>