[IMP] iot: offline pairing with token
As we removed the pairing token to ease the pairing process by automating it, we can't pair an IoT Box with
a database there is no access to the internet (automated pairing depend on odoo.com).

We then added back the pairing token, generated by the database, to provide to the IoT Box Homepage.

closes odoo/enterprise#87451

Task: 4844687
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>