[IMP] account_accountant: improve recognition of multiple invoices
In this commit:

Improved recognition of multiple invoices in bank statement line when:
- Bank statement line and invoices have the same `partner_id`, the amount
  matches their combined total, and `payment_ref` contains all invoice
  references
- When any 2 of the 3 criteria are satisfied

Also, we enhance the reconciliation logic by adding a check to match
the payment memo against the bank statement line label. If a part of the memo is
found exactly in the label and only one such match exists, the system proceeds
to reconcile it based on a second matching criterion such as amount and/or
partner.

Also enable full-text search for invoice references within `payment_ref`
(e.g., "INV/2025/00027" found in "Payment for INV/2025/00027").

closes odoo/enterprise#87676

Task: 4797388, 4791517, 4852948
X-original-commit: c0f8c5d17b9d78b640d744d8bfb6a2bb4002776a
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Co-authored-by: hatr-odoo <hatr@odoo.com>
Co-authored-by: igbe-odoo <igbe@odoo.com>