[FIX] account_accountant: reco model lang
The name of the reconcile models displayed in the widget was sometimes empty due
to a mismatch in user language. Specifically, if a reconcile model was created
in one language (e.g., en_US) and the user switched to a different language, the
model name would not be found and thus appear empty.
This led to a props validation error in the widget.

task-4863754

closes odoo/enterprise#87612

X-original-commit: a6292224fd5de735d9ebfd62108b638be0d20525
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>