[FIX] l10n_lu_reports, l10n_ca_reports: fix bad migration of cross_report
Some aggregation expressions used to have 'cross_report' as subformula. Though, it was useless (since the aggregation only uses terms from the same report), and the subformula was removed from the data file without explicitly resetting it to False. This became a problem in 18.3, because the cross_report syntax changes. Because of that, a migrated report failed to open, since it still was using the old syntax on that expression. We fix that by explicitly emptying the subformula.

This was done in https://github.com/odoo/odoo/commit/bbba54a08c56e0a4d040fc8035996d9d7cafedad for the monthly Luxembourgish tax report ; but some other occurrences needed to be fixed.

closes odoo/enterprise#87558

X-original-commit: 13050da4dbeabf79fc48503f0ec2398e49f076e7
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>