[FIX] documents_spreadsheet: Fix failing multipage tour
The multipage tour started failing after an update of the lib
o-spreadsheet. While no significant change seems to explain the failing,
it turns out that the tour makes LOTS of useless steps and the actual
feature tested doesn't even require to load a spreadsheet.

This revision simplifies the tour as much as possible.

Fixes runbot error 226565

closes odoo/enterprise#87493

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>