[FIX] esg: make company non required on factors
Having a mandatory company on factors comes with multiple problems.
The main one is that users rely on factors imported from huge databases, that would therefore have to be imported once per company.
The current dabatase model is incidentally built around factors being available in all companies, which means one cannot actually import the factors multiple times.
Another is that users could see sources (which are not company dependent) that are seemingly linked to no factors because they are in other companies, which can be confusing.

To alleviate those problems, this commit makes the field non required.
Note that this will only take effect for new databases or after an upgrade.

closes odoo/enterprise#87557

X-original-commit: bb8bf7af98c821129a8398e520caf8cfe616b0aa
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>