[FIX] voip: handle contacts with falsy/undefined name.
An error was arising when making the following steps:
1. create a company contact with a proper name.
2. create an invoice address for that company through the company form
   view.
3. set a phone number for the invoice.
4. leave the `name` empty.
5. open the contacts tab on VoIP.

This happened because the code was trying to access the first
charachater of a falsy string which is `voipName` in this case. A
solution to this was to guard falsy names of contacts and add a section
for contacts with empty string. This section will have a title of "#"
and will appear at the end of the list.

closes odoo/enterprise#87675

X-original-commit: e98cc48b0725dd9bb1308317010e5e3cadab4847
Signed-off-by: Yuchen Huang (yhu) <yhu@odoo.com>
Signed-off-by: Mohammed Basioni (basm) <basm@odoo.com>