[FIX] documents: set focus on document when using select all
Before this commit, when selecting all (ctrl-a) in kanban or list views the
first document is shown in the detail panel but the focus is not set on it.
This commit fix this issue, when no document is focused and we select all the
focus is set on the first document, if a document is already selected when
selecting all the focus will still be that document.

Task-4813624

closes odoo/enterprise#88852

X-original-commit: 1321599504ee4ed281f82ca7ccd359b3a235c72e
Related: odoo/odoo#216443
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>