[IMP] mrp_workorder: estimate cost of operation
The cost of a MO is currently based on the actual time.
In the case a worker worked 3 hours but only logged in Odoo to
register a serial number, the actual time would only be a few
seconds.

For this reason, we are adding the possibility to compute the cost
based on the duration_expected of the workorder.

The conditions to use the duration_expected are:
- the workorder is linked to an existing operation
- the cost_mode of the operation is set to 'estimated'
- there's a duration_expected on the workorder

This affects the MO overview and the valuation layers.

task 4850419

closes odoo/enterprise#87522

Related: odoo/odoo#213762
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>