[FIX] l10n_it_riba: fix rounding issue in batch payment export
Rounding errors may occur when using `int()` on float values.

Steps to reproduce:
- Create an invoice with a total amount of 256.46
- Create a RIBA payment for this invoice
- Add this payment to a batch payment
- Export the batch payment file

In the exported file, the amount should be 25646, but it is incorrectly
25645 in two places.

This happens because in Python:
`int(256.46 * 100) → int(25645.999999...) → 25645`

To fix this, we use `float_round()` to round correctly the float before
applying `int()` to convert the value to an integer.
This ensures accurate rounding for exported payment amounts.

opw-4858205

closes odoo/enterprise#87770

X-original-commit: e442f56013c1a731f0dc456cc44e7a19bbc3c5e8
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>