[IMP] account_accountant, account_bank_statement_import: statement attachment
Before this commit, the statement attachment was created only when doing an
import but not when creating the statement manually. This commit will move the
logic to create the attachment into the create of account bank statement

When manually putting a pdf into the statement, we don't create the attachment
because it's already added to the statement

For demo data, the connection to wkhtml2pdf is unreachable and it makes a
warning on runbot. The context key is also used for that.

task-4866269

closes odoo/enterprise#88293

X-original-commit: 0d4132f60144b397b44ba9fe25b4c08dd79a40c5
Related: odoo/odoo#215474
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>