[FIX] pos_blackbox_be: did not send ACK after status message
We did not send an `ACK` after asking for blackbox status, leading in BlackBox sending it again.
This lead to situations where we sent requests to the BlackBox, and received the status instead of
the actual response.
In addition, the repeated status responses were missing the leading `ACK`, resulting in our log:
`sent 1 NACKS without receiving response, giving up.`

More, if we sent a command too soon to the BlackBox (before it finished instanciating), we were receiving its
`certified ref` instead of the actual response, due to the `_push_status` method called from
the `run` method of the BlackBox Driver parent class.
This `run` method is now overridden by the BlackBox driver

closes odoo/enterprise#87988

Task: 4750435
X-original-commit: c4c605701ac5ee606ac0d3b3a080b9739a7c7e8c
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>