[FIX] l10n_br_edi_pos: stop invoicing automatically on company customers
Since [^1] company customers are always invoiced by default. To Invoice
should always be false for Brazil POS orders which is why the button on
the payment screen was hidden, however, there is now no way to uncheck
it.

Fix:
Set "To Invoice" to false always if the company's fiscal country is
Brazil.

[^1]: https://github.com/odoo/odoo/pull/202515

opw-4861672

closes odoo/enterprise#87691

X-original-commit: 95a0a67d181aac29a611266a0324293f97d91da4
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Andrew Gavgavian (andg) <andg@odoo.com>