[IMP] planning: allow to create multiple shifts based on template
Before this commit, when the user selects multiple gantt cells to create
a new shift and select a shift template, a shift will be created for the
whole period selected. In that case, the user could prefer to create one
shift per day to correctly follow the shift template.

This commit will generate one shift per day when the user selects
multiple days in the gantt view and selects a shift template and save.

task-4801348

closes odoo/enterprise#87711

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>