[IMP] ai: integrate & improve AI draft with AI agents
In this commit we are integrating the AI draft functionality introduced
in 18.3 with the AI agents also introduced in 18.3.

We heavily updated the ai.composer model in order to act more like AI
draft configurations that are editable and extensible by the user. We
introduced the `interface_key`, 'ai_agent', 'focused_models', and the
'available_prompts' fields.

We also introduced a new model, the `ai.prompt` model, which holds str
of the available prompts that users can click on when opening a AI draft
chat.

Finally, we made some UI changes to the chat windows of AI chats - they
now have configurable prompt buttons that the users can immediately
click when opening the chat. We also moved the chatter button from the
chatter to the form view header.

(Also we edited some tests, and some legacy qunit tests, because the
`aiChatLauncher` service needed to be added since it's not being used
by the `form_controller`)

Task-4794888

closes odoo/enterprise#91331

Related: odoo/odoo#221219
Related: odoo/upgrade#8143
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>