[IMP] sign: adding read-only option to fields
This commit introduces a new read-only option to the sign items. This
option allows the user to add fields that are non-editable when signing.
The value that is displayed is either the defined placeholder or the
auto-value for the field.

This commit also introduce the possibility to resize the input of the
placeholder for the sign items.

Regarding the technical choices, this commit adds a constraint to the
`sign.item` cannot be read-only and required at the same time. This is
enforced both from the front-end by disabling inputs, and in the
back-end by a hard constraint.

task-4606205

closes odoo/enterprise#81357

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>