[IMP] account_asset: add properties button on assets and asset model
This commit introduces two improvements to the 'Add Properties' button
in the cog menu of the Asset model:
1. Hides the button if the record has state = 'model'. Previously, the
button was available but no action was triggered.
2. Raises a better warning if the button 'Add Properties' is clicked in
an asset without an asset model associated. Previously, only a warning
with the unclear message "You need edit access on the parent doument
to update these property fields" was shown.

task-4822091

closes odoo/enterprise#87813

X-original-commit: 249e8ed4324565692e418c0afd22e04e86704144
Related: odoo/odoo#214543
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>