[FIX] hr_expense_extract: Fix Bad OCR computation
Fixes a bug where the OCR would sometimes put a foreign currency
on an expense with a product having a cost.
Making it impossible to switch back to the company currency
(because the currency cannot be changed).

This also allows the user to change the total amount after the OCR pass,
so it can be corrected if needed

Steps to reproduce:
- Setup accounting and expense, in a non-bosnian company
- OCR a google map screenshot or simulate it, the selected product
 must have a standard_price setup (flat-rate flow)

Problem:
- OCR changes the currency to Convertible Marks (kms)
- Product having a cost prevents the user from changing
  the currency back.
- Sad employee can't be reimbursed

task-4873236

closes odoo/enterprise#101887

X-original-commit: 0ae813c72e66881fc4aa4b4cce9e4ef30942720b
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>