[FIX] sale_amazon: support dpd and hermes in GB as carrier
For DPD and Hermes (not the others), Sendcloud adds an unnecessary ISO code.

In order to keep it should it be necessary (e.g. rml_gb), we map them to the
Amazon const instead of stripping the code.

While testing, it was also found that royal mail seems to have been renamed
and was thus not working.

closes odoo/enterprise#89054

X-original-commit: 2cc1bce28b9ffd7c8af8e55bb15f9c3eeea89c72
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>