[FIX] sign: auto scrool to next item was stopping early
Before this commit, in Chrome, when the first item was placed on second
page or later, clicking on "Click to start" to automatically scroll to
the first sign item would stop earlier. As a result, the user would have
to click again on the navigator to focus it and to fill the automatic
value.

task-id-4854650

closes odoo/enterprise#87998

X-original-commit: d62c7174c7e2b9dfc398ead863930758b9eb1c2c
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>