[FIX] sale_subscription: fix main subscription qty change after upsell
Steps to reproduce:

- Install sale_subscription
- Create a recurring product (P1) and a non-recurring product (P2)
- Create a Sale Order (S1) with P1 (quantity > 0) and confirm it
- Create an upsell of S1 (new child order S2)
- In S2, edit the P1 line, and select P2
- Set quantity > 0 for P2
- Confirm S2

Issue:

After confirming the upsell order (S2), the quantity of P1 on the
original subscription (S1) is incorrectly increased by the quantity of
P2 from S2. The _compute_parent_line_id method did not reset the
parent_line_id when the product was not a subscription Product,
which caused it to incorrectly update the subscription

Solution:

Reset parent_line_id to False if the product is not a subscription
product to prevent incorrect subscription updates.

opw-4638955

closes odoo/enterprise#90964

X-original-commit: a84ce70e866414f91363b724f8b8d021f10bd76f
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Hiral Bhavsar (hbh) <hbh@odoo.com>