[IMP] l10n_co_{dian,edi,edi_mandate}: support AIU products
AIU (Administration, Contingencies, and Profit) is used to keep track of
costs and sales tax in service contracts like construction or
engineering. Any indirect admin expenses, unforseen costs, or profits
for the contractor is included in this data.

As this is a common requirement the three products are created as data
records and will be available to all customers using our EDI service.
The UBL specifications state that for AIU invoices we must include a
specific string in the note section of the e-invoice on all product
lines.

To accomplish this, the operation type field had to be made into a
stored compute field that uses the pre-established compute as a default
value based on type of move.

Due to this change, to keep the default behaviour the mandate module has
changed from setting the operation on a compute to instead show a
warning if the right operation type is not properly set as it is cleaner
and does not recompute the value every time a line is edited.

task-4611282

closes odoo/enterprise#86737

Related: odoo/upgrade#7873
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>